service/requirements.txt
web.py
requests
google-auth
firebase-admin
google-cloud-aiplatform